+++ Src/Tools/Entity.h
@@ -20,9 +20,9 @@ public:
   public:
     void delegateEntity(Entity& entity)
     {
-      quint32 id = entity.getId();
+      quint64 id = entity.getId();
       EntityTable& table = tables[entity.getType()];
-      QHash<quint32, Entity*>::iterator it = table.entities.find(id);
+      QHash<quint64, Entity*>::iterator it = table.entities.find(id);
       if(it == table.entities.end())
       {
         table.entities.insert(id, &entity);
@@ -41,9 +41,9 @@ public:
 
     void delegateEntity(Entity& entity, Entity& replacedEntity)
     {
-      quint32 replacedId = replacedEntity.getId();
+      quint64 replacedId = replacedEntity.getId();
       EntityTable& replacedTable = tables[replacedEntity.getType()];
-      QHash<quint32, Entity*>::iterator itReplaced = replacedTable.entities.find(replacedId);
+      QHash<quint64, Entity*>::iterator itReplaced = replacedTable.entities.find(replacedId);
       if(itReplaced != replacedTable.entities.end())
       {
         replacedTable.entities.erase(itReplaced);
@@ -51,9 +51,9 @@ public:
           (*i)->removedEntity(replacedEntity, entity);
       }
 
-      quint32 newId = entity.getId();
+      quint64 newId = entity.getId();
       EntityTable& newTable = tables[entity.getType()];
-      QHash<quint32, Entity*>::iterator it = newTable.entities.find(newId);
+      QHash<quint64, Entity*>::iterator it = newTable.entities.find(newId);
       if(it == newTable.entities.end())
       { // move replacing entity to new entity
         newTable.entities.insert(newId, &entity);
@@ -83,10 +83,10 @@ public:
           (*i)->updatedEntitiy(entity, entity);
     }
   
-    void removeEntity(quint32 type, quint32 id)
+    void removeEntity(quint32 type, quint64 id)
     {
       EntityTable& table = tables[type];
-      QHash<quint32, Entity*>::iterator it = table.entities.find(id);
+      QHash<quint64, Entity*>::iterator it = table.entities.find(id);
       if(it != table.entities.end())
       {
         Entity* entity = *it;
@@ -97,12 +97,12 @@ public:
       }
     }
 
-    template <typename E> void removeEntity(E eType, quint32 id)
+    template <typename E> void removeEntity(E eType, quint64 id)
     {
       removeEntity((quint32)eType, id);
     }
 
-    template <typename C> void removeEntity(quint32 id)
+    template <typename C> void removeEntity(quint64 id)
     {
       removeEntity((quint32)C::eType, id);
     }
@@ -112,7 +112,7 @@ public:
       EntityTable& table = tables[type];
       if(!table.entities.isEmpty())
       {
-        QHash<quint32, Entity*> entities;
+        QHash<quint64, Entity*> entities;
         entities.swap(table.entities);
         for(QSet<Listener*>::iterator i = table.listeners.begin(), end = table.listeners.end(); i != end; ++i)
           (*i)->removedAll(type);
@@ -125,19 +125,19 @@ public:
       removeAll((quint32)C::eType);
     }
 
-    Entity* getEntity(quint32 type, quint32 id) const
+    Entity* getEntity(quint32 type, quint64 id) const
     {
       QHash<quint32, EntityTable>::ConstIterator itTable = tables.find(type);
       if(itTable == tables.end())
         return 0;
       const EntityTable& table = *itTable;
-      QHash<quint32, Entity*>::ConstIterator it = table.entities.find(id);
+      QHash<quint64, Entity*>::ConstIterator it = table.entities.find(id);
       if(it == table.entities.end())
         return 0;
       return *it;
     }
 
-    template<class C> C* getEntity(quint32 id) const
+    template<class C> C* getEntity(quint64 id) const
     {
       return dynamic_cast<C*>(getEntity((quint32)C::eType, id));
     }
@@ -147,18 +147,18 @@ public:
       QHash<quint32, EntityTable>::ConstIterator itTable = tables.find((quint32)C::eType);
       if(itTable != tables.end())
       {
-        const QHash<quint32, Entity*>& hashMap = itTable->entities;
-        for (QHash<quint32, Entity*>::ConstIterator i = hashMap.begin(), end = hashMap.end(); i != end; ++i)
+        const QHash<quint64, Entity*>& hashMap = itTable->entities;
+        for (QHash<quint64, Entity*>::ConstIterator i = hashMap.begin(), end = hashMap.end(); i != end; ++i)
           entities.append(dynamic_cast<C*>(*i));
       }
     }
 
-    template<class C> quint32 getNewEntityId()
+    template<class C> quint64 getNewEntityId()
     {
       EntityTable& table = tables[(quint32)C::eType];
       if(table.entities.isEmpty())
         return 1;
-      quint32 id = (--table.entities.end()).value()->getId();
+      quint64 id = (--table.entities.end()).value()->getId();
       while(table.entities.contains(id))
         id += qrand() % 10;
       return id;
@@ -189,13 +189,13 @@ public:
   private:
     struct EntityTable
     {
-      QHash<quint32, Entity*> entities;
+      QHash<quint64, Entity*> entities;
       QSet<Listener*> listeners;
 
       ~EntityTable()
       {
         //qDeleteAll(entities);
-        for(QHash<quint32, Entity*>::Iterator i = entities.begin(), end = entities.end(); i != end; ++i)
+        for(QHash<quint64, Entity*>::Iterator i = entities.begin(), end = entities.end(); i != end; ++i)
         {
           Entity* e = *i;
           delete e;
@@ -207,13 +207,13 @@ public:
 
 
 public:
-  template <typename E> Entity(E type, quint32 id) : type((quint32)type), id(id) {}
+  template <typename E> Entity(E type, quint64 id) : type((quint32)type), id(id) {}
   virtual ~Entity() {}
 
   quint32 getType() const {return type;}
-  quint32 getId() const {return id;}
+  quint64 getId() const {return id;}
 
 private:
   quint32 type;
-  quint32 id;
+  quint64 id;
 };
